Build: Adds a bunch. Requires more building

+++ hw1/src/hw1/Triage.java
@@ -7,7 +7,7 @@ public class Triage extends CareArea {
 
     public static final int NUM_OF_SERVERS = 1;
 
-    public Triage(int expectedAvgInterArrivalTimePerPatient, int expectedAvgServiceTimePerPatient, int maxQueueLength) {
+    public Triage(double expectedAvgInterArrivalTimePerPatient, double expectedAvgServiceTimePerPatient, int maxQueueLength) {
         super(expectedAvgInterArrivalTimePerPatient, 
               expectedAvgServiceTimePerPatient, 
               maxQueueLength,
